Added NoFeatureBaseStructure
